refactor: Move statements logic to own module
* Move gucHeader decoding to Statements
